Move the "descending first?" boolean into the field definition
It's a bit cleaner to put it in the field definition instead of only
closing over it during registration, and also allows the toggleSort()
method to be called without knowledge of the field's initial
descendingFirst value (such as from auxiliary template elements).

This also makes the fields property a Map instead of an array for easier
retrieval.